requirements.txt
attrs==21.2.0
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.9
coverage==6.2
cx-Oracle==8.2.1
flake8==4.0.1
google-api-core==1.25.0
google-auth==1.22.1
google-auth-oauthlib==0.4.2
google-cloud-bigquery==2.3.1
google-cloud-bigquery-storage==2.0.1
google-cloud-core==1.5.0
google-cloud-secret-manager==2.2.0
google-cloud-storage==1.35.0
google-crc32c==1.3.0
google-pasta==0.2.0
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
graphframes-dev==0.7.0
grpc-google-iam-v1==0.12.3
grpcio==1.43.0
idna==3.3
importlib-metadata==4.2.0
iniconfig==1.1.1
jellyfish==0.8.9
joblib==1.1.0
libcst==0.3.23
mccabe==0.6.1
mypy-extensions==0.4.3
nose==1.3.3
numpy==1.19.5
oauthlib==3.1.1
packaging==21.3
pandas==1.1.5
pluggy==1.0.0
proto-plus==1.19.8
protobuf==3.19.1
py==1.11.0
py4j==0.10.9
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.6
pyspark==3.1.2
pytest==6.2.5
pytest-cov==3.0.0
pytest-runner==5.3.1
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
recordlinkage==0.14
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
scikit-learn==0.24.2
scipy==1.5.4
six==1.16.0
smartystreets-python-sdk==4.9.0
threadpoolctl==3.0.0
toml==0.10.2
tomli==1.2.3
typing-inspect==0.7.1
typing_extensions==4.0.1
urllib3==1.26.7
XlsxWriter==3.0.1
zipp==3.6.0
